+++ src/dist_s1/packaging.py
@@ -164,7 +164,7 @@ def package_disturbance_tifs_no_confirmation(run_config: RunConfigData) -> None:
         ),
     ]
 
-    tags = run_config.get_public_attributes()
+    tags = run_config.get_public_attributes(include_algo_config_params=True)
     tags['version'] = dist_s1.__version__
     tags = update_tags_with_opera_ids(tags)
     tags = update_tag_types(tags)