+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `base_date` is now `base_date_for_confirmation`.
 - Remove algorithm assignments in localization workflows.
 - Organized `AlgoConfigData` and `RunConfigData` into seperate files
+  - `AlgoConfigData` is an attribute of the `RunConfigData`
+  - Retrieval (and serialization) of the Algorithm Parameters can still be obtained via `get_public_attributes`.
 - Put data/path utils into `data_utils.py`.
 - Variables for confirmation processing (use snake case wherever possible)
 - Now have an output product without confirmation and a product with confirmation (for provenance)
@@ -45,9 +47,12 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - Exposed parameter `percent_reset_thresh`, it will apply reset if `percent` below threshold. 
   - Exposed parameter `nocount_reset_thresh`, it will apply reset if `prevnocount` is above threshold.   
 - Confirmation CLI
+  - Ensuring additional profile keys for in default GTiff are not present.
 
 ## Fixed
 - All the tests with the updates above.
+- Correct loading of algo_config.yml in `prep` steps.
+  - Only `from_yml` loads the `algo_config` correctly, but when it is assigned in the `prep` workflows (i.e. the attribute `algo_config` is assigned, this yml is not correctly loaded.
 
 
 ## [1.0.1] - 2025-06-05